renamed ssid to jwt and renamed sessionController to tokenController. websocket connection now does real authentication with jwt cookie
