+++ src/server/routes/userRouter.js
@@ -1,22 +1,22 @@
 const express = require('express');
 const userController = require('../controllers/userController')
-const sessionController = require('../controllers/sessionController')
+const sessionController = require('../controllers/tokenController')
 
 const router = express.Router();
 
-router.post('/signup', 
+router.post('/signup',
   userController.createUser,
-  sessionController.setSSIDCookieSignUp,
+  sessionController.setJWTCookieSignUp,
   userController.setIsOnlineTrue,
   (req, res) => {
     res.locals.userExists ? res.status(200).json('username exists') :
-    res.status(200).json(res.locals.userInfo)
+      res.status(200).json(res.locals.userInfo)
   }
 );
 
 router.post('/login',
   userController.verifyUser,
-  sessionController.setSSIDCookieLogin,
+  sessionController.setJWTCookieLogin,
   userController.setIsOnlineTrue,
   (req, res) => {
     res.locals.authenticated ? res.status(200).json(res.locals.userObj) : res.status(200).json('incorrect password or username');
@@ -24,7 +24,7 @@ router.post('/login',
 );
 
 router.post('/logout',
-  sessionController.removeSSIDCookie,
+  sessionController.removeJWTCookie,
   userController.setIsOnlineFalse,
   (req, res) => {
     res.status(200).json('Logged out');
